pyproject.toml
[project]
name = "data-analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "splicid",email = "luis1abreu11@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "pandas (<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "scikit-surprise (>=1.1.4,<2.0.0)",
    "numpy (<2.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
